Added exercises 22-27
Did a lot of work on the section 2 prime test questions. Note that Chez
Scheme doesn't include a runtime marker, so I can't actually test the
time efficiency of these programs. The numbers are so quick to calculate
nowadays that I'd have to find larger numbers to replace them with to
notice a change in the runtimes anyway.