Renamed exercises in ch1 sn2
More renaming. I'm thinking of leaving in the other, non-exercise
non-notetaking files, but with a different prefix to distinguish them. I
could add them to notes.md, but then I'd have to mess with the structure
of my notes a lot more, and I'd like to keep the code in my notes
relatively light just to demonstrate points as necessary, if at all.