Added exercises 1.33, 1.34, and 1.35
The Spring 2021 semester is over, which means I can spend more time
working on learning Scheme and other functional languages. In the past
week I have updated my paper notes and read ahead to chapter 2, section
3.

Exercises 1.33-1.35 are about the accumulate function. This is called
fold in Haskell. It's one component of stream processing, which is
discussed further in chapter 2. The function given isn't even the most
general function you can use for this, but that's not important atm.