Added basic interpreter skeleton
Added the code from section 4.1.1 through 4.1.4 of the book. This is a
big jump from working on chapter 2, but I've been growing a bit
frustrated with the mathematically dense code tht doesn't reveal
interesting insights into Scheme's data model beyond simple arithmetic
representations of things like line segments.

This code doesn't actually work out of the box, and there are several
things that I just straight-up forgot to include. Much restructuring is
warranted in getting it working, and it may be worthy of its own GitHub
repo for completion.