Added N-queens problem, unsolved
Here is the initial sample code for the N-queens problem on page 125 of
the book, Exercise 2.42. I get the feeling the authors have a very
specific solution in mind, but there may be other methods of tackling
this problem, so I have saved the problem here to be modified later.