Added 2.43
It's crazy how such a small change in Louis's code can have such a
profound impact on the search space of a recursive algorithm. This
should serve as a reminder that not all helper functions can be treated
as running in O(1) time. Also, I cleaned up Exercise 2.42 a bit because
I'm proud of how it turned out and wanted to show it to people in the
future.