Added lambda calculus stuff
Did a lot of work implementing the Lambda Calculus! I find the use of
the lambda symbol for anonymous procedures really neat, and I hope to do
it in the future with other programs.

TODO: finish the Python version of the same code, and add more examples
of uses and applications for the lambda calculus. Clean up comments
explaining everything, and get ready for presenting!