Initial stages of lazy evaluation complete
Work heavily under progress on the lambda calculus operators constructed
in Scheme. I am finally on the path to resolving my issues with eager
evaluation of predicates and their consequents/alternatives, but I have
yet to generalize this to all if-then-else statements. Of course the
true solution is to write an evaluator myself that evaluates EVERYTHING
in normal order. See chapter 4, section 2 of SICP.