Started on 2.65
Added a (bad) representation of union-set for trees. The output doesn't
work properly since it adds multiple instances of a single number to the
tree, but otherwise it's headed in the right direction. This is what I
get for not following the book's hint to utilize tree->list and
list->tree.

TODO: rewrite union-set and intersection-set to use lists also.